feat: add MicroserviceEntry type
