+++ Server/Gateway/.env
@@ -1,1 +1,2 @@
-TEST_HOST=localhost:8081
+CONTENT_HOST=localhost:8081
+FEEDBACK_HOST=localhost:8082
