+++ Server/Gateway/Sources/App/Controllers/GatewayController.swift
@@ -7,27 +7,19 @@
 
 import Vapor
 
-extension HTTPMethod {
-    public static var redirectableMethods: [HTTPMethod] {
-        return [GET, PUT, POST, DELETE]
-    }
-}
-
-extension RoutesBuilder {
-    func use<Response: ResponseEncodable>(handler: @escaping (Request) throws -> Response) {
-        HTTPMethod.redirectableMethods.forEach {
-            on($0, "*", use: handler)
-        }
-    }
-}
-
 final class GatewayController: RouteCollection {
+    
+    private let microservices: [MicroserviceEntry] = [.init(path: "content", host: "CONTENT_HOST"),
+                                                      .init(path: "feedback", host: "FEEDBACK_HOST")]
+    
     func boot(routes: RoutesBuilder) throws {
-        routes.grouped("test").use(handler: handle)
+        for entry in microservices {
+            routes.grouped(entry.path).use(handler: handle)
+        }
     }
     
     func handle(_ req: Request) throws -> String {
-        return "Hello, world!"
+        return "Welcome to \(req.url.path)!"
     }
     
     func redirect(_ req: Request) throws -> EventLoopFuture<ClientResponse> {
@@ -42,7 +34,7 @@ final class GatewayController: RouteCollection {
         let client = req.client
         let uri = URI(string: host + req.url.string)
         var headers = req.headers
-        headers.replaceOrAdd(name: "host", value: host)
+        headers.replaceOrAdd(name: .host, value: host)
         let request = ClientRequest(method: req.method, url: uri, headers: headers, body: req.body.data)
         return client.send(request)
     }
