+++ Server/Gateway/Sources/App/Controllers/MicroserviceEntry.swift
@@ -1,0 +1,18 @@
+//
+//  File.swift
+//  
+//
+//  Created by Aidar Nugmanov on 4/25/20.
+//
+
+import Vapor
+
+final class MicroserviceEntry {
+    let path: PathComponent
+    let host: String?
+    
+    init(path: String, host: String) {
+        self.path = PathComponent.constant(path)
+        self.host = Environment.get(host)
+    }
+}
