+++ Server/Gateway/Sources/App/Extensions/RoutesBuilder+Use.swift
@@ -1,0 +1,16 @@
+//
+//  RoutesBuilder+Use.swift
+//  
+//
+//  Created by Aidar Nugmanov on 4/25/20.
+//
+
+import Vapor
+
+extension RoutesBuilder {
+    func use<Response: ResponseEncodable>(handler: @escaping (Request) throws -> Response) {
+        HTTPMethod.redirectableMethods.forEach {
+            on($0, PathComponent.anything, use: handler)
+        }
+    }
+}
